categories/web/web-easy-Emergency-access/give/service_provider/requirements.txt
Flask==3.1.0
lxml==5.3.2
signxml==4.0.3
defusedxml==0.7.1
xmlsec==1.3.15
python-dotenv==1.0.1
gunicorn==21.2.0
pytz==2022.7

